Configure code to support different architectures
