+++ .gitignore
@@ -22,6 +22,7 @@ tmp/
 # Models (weights)
 results/
 models/
+!ml/models
 notebooks/tmp/*.hdf5
 
 # Config
