+++ ml/utils.py
@@ -1,0 +1,38 @@
+"""Custom metrics, callbacks, and plots.
+"""
+import keras
+import sklearn
+from keras import backend as K
+
+
+def true_positives(y_true, y_pred):
+    return K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
+
+
+def false_negatives(y_true, y_pred):
+    return K.sum(K.round(K.clip(y_true * (1 - y_pred), 0, 1)))
+
+
+def sensitivity(y_true, y_pred):
+    true_positives = K.sum(K.round(K.clip(y_true * y_pred, 0, 1)))
+    possible_positives = K.sum(K.round(K.clip(y_true, 0, 1)))
+    return true_positives / (possible_positives + K.epsilon())
+
+
+def specificity(y_true, y_pred):
+    true_negatives = K.sum(K.round(K.clip((1 - y_true) * (1 - y_pred), 0, 1)))
+    possible_negatives = K.sum(K.round(K.clip(1 - y_true, 0, 1)))
+    return true_negatives / (possible_negatives + K.epsilon())
+
+
+class AucCallback(keras.callbacks.Callback):
+
+    def __init__(self, x_valid_standardized, y_valid):
+        super().__init__()
+        self.x_valid_standardized = x_valid_standardized
+        self.y_valid = y_valid
+
+    def on_epoch_end(self, epoch, logs=None):
+        y_pred = self.model.predict(self.x_valid_standardized)
+        score = sklearn.metrics.roc_auc_score(self.y_valid, y_pred)
+        print(f'\nval_auc: {score}')
